adding controllers, moving classes, deleting don't needed classes
