+++ Application/Application/src/main/java/com/example/Application/repository/MaterialRepository.java
@@ -1,6 +1,6 @@
 package com.example.Application.repository;
 
-import com.example.Application.core.Material;
+import com.example.Application.model.Material;
 
 import java.util.List;
