+++ Application/Application/src/main/java/com/example/Application/repository/datajpa/CrudTicketRepository.java
@@ -1,8 +1,7 @@
 package com.example.Application.repository.datajpa;
 
-import com.example.Application.core.Material;
-import com.example.Application.core.Ticket;
-import com.example.Application.model.User;
+import com.example.Application.model.Material;
+import com.example.Application.model.Ticket;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
