+++ Application/Application/src/main/java/com/example/Application/repository/datajpa/DataJpaMaterialRepository.java
@@ -1,6 +1,6 @@
 package com.example.Application.repository.datajpa;
 
-import com.example.Application.core.Material;
+import com.example.Application.model.Material;
 import com.example.Application.repository.MaterialRepository;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Repository;
