+++ Application/Application/src/main/java/com/example/Application/web/user/AbstractUserController.java
@@ -1,0 +1,26 @@
+package com.example.Application.web.user;
+
+import com.example.Application.model.User;
+import com.example.Application.service.UserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import java.util.List;
+
+public class AbstractUserController {
+    protected final Logger log = LoggerFactory.getLogger(getClass());
+
+    @Autowired
+    private UserService service;
+
+    public List<User> getAll() {
+        log.info("getAll");
+        return service.getAll();
+    }
+
+    public User get(int id) {
+        log.info("get {}", id);
+        return service.get(id);
+    }
+}
