+++ Application/Application/src/main/java/com/example/Application/web/user/AdminUIController.java
@@ -1,0 +1,12 @@
+package com.example.Application.web.user;
+
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+
+@RestController
+@RequestMapping(value = "/admin/users", produces = MediaType.APPLICATION_JSON_VALUE)
+public class AdminUIController extends AbstractUserController {
+
+}
