+++ Application/Application/src/test/java/com/example/Application/core/MaterialTest.java
@@ -2,15 +2,13 @@ package com.example.Application.core;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import com.example.Application.model.Material;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.stereotype.Component;
-import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.context.junit4.SpringRunner;
 
 //@ExtendWith(SpringExtension.class)
